Split profile page into subroutes
