+++ src/components/views/LinkAccountsView.vue
@@ -1,0 +1,9 @@
+<template>
+  <h2>Link Accounts</h2>
+  <span>Make logging in easy by linking your accounts.</span>
+  <div class="button-container">
+    <PvButton>Log in with Clever</PvButton>
+    <PvButton>Log in with Google</PvButton>
+    <PvButton>Log in with ClassLink</PvButton>
+  </div>
+</template>
